feat: Added `sherlock` command to OSINT
I opted to remove the `sherlock` command from the `opsec` section. Simply put, it does not fit the vision I have, it's not very R6 related. It was originally intended to amplify the strength of the lookup command, but people are purchasing the tool solely for that purpose. Accordingly, I am moving it to OSINT.