feat: Drastically cleaned up the help sections and menus
I was losing faith partway through, but I'm so much happier with the results :))). I genuinely never want to think about async recursion again - it was entirely a skill issue, but it's amazing how much easier synchronus recursion is. With the past commits, the bot's speed should increase dramatically, and it should be able to better concurrently handle requests for the ever-growing number of users!