feat!: Implemented an automatic MOSS file check feature and player `recon` command
The `mosscheck` command is for locating additional accounts and pre-screening MOSS files for any suspicious behavior.
The `recon` command is specifically for investigating whether a player is boosted, a boostee, or previously banned.

Developer Aside: Requires `OPENAI_API_KEY` to be set to a valid OpenAI legacy or project key.