feat, bugfix!: Began overhauling Discord functionality
Changes:
The primary goal is for this application to function as a web API first, and Discord bot second. This is accomplished by making a more generic `Sendable` enum, with Discord as one of the options. By making a genric set of functions that can be called no matter the frontend, I don't need to create an entirely seperate API - it can be automatically generated. Difficult, sure, but will save me a lot of work in the long run. This commit does *not* fully implement this, only creates the nessecary framework to be worked on in the future. Future commits will implement the API `Sendable`.

Bugfixes:
Fixed multiple commands simply not responding if there were missing arguments. Disabled `mosscheck` command as it doesn't present the information in a manner I like. Instead, I will probably make it either text-document only for Discord, or limit it to only being usable on the web frontend.

Developer Note:
You will need to add a PORT environment variable set to a port of your choice to prevent build failure.