feat: Added incredibly strong edit and finalization race condition check system
Discord's edit API returns OK even despite the message not being edited, which is a problem. This means, in cases where I am processing a command in parallel segments, otherwise sequential edits enter a race condition and appear in an arbitrary order.

To solve this, I implemented a checking system that atomically confirms the edits are happening sequentially, and that the body was not tampered with within interim.

Pretty cool stuff, atomic generics are nice - the control over ordering you get compared to an atomically reference counted pointer template is slick.

*This is honestly a change made to satiate my interest in seeing if I could, and I'm usually glad I do these things in the end. Sure, with <200 users, this level of hyper-optimization might be a little needless. It doesn't make it any less cool to understand how to do things like this though.*